Factor out UsageTrackingCommand
