+++ src/GitHub.InlineReviews/Services/PullRequestStatusBarManager.cs
@@ -5,14 +5,12 @@ using System.Windows.Controls;
 using System.Windows.Controls.Primitives;
 using System.ComponentModel.Composition;
 using System.Reactive.Linq;
-using System.Linq.Expressions;
 using GitHub.Commands;
 using GitHub.InlineReviews.Views;
 using GitHub.InlineReviews.ViewModels;
 using GitHub.Services;
 using GitHub.Models;
 using GitHub.Logging;
-using GitHub.Extensions;
 using Serilog;
 using ReactiveUI;
 
@@ -45,10 +43,10 @@ namespace GitHub.InlineReviews.Services
             Lazy<IPullRequestSessionManager> pullRequestSessionManager,
             Lazy<ITeamExplorerContext> teamExplorerContext)
         {
-            this.openPullRequestsCommand = new UsageTrackingCommand(openPullRequestsCommand,
-                usageTracker, x => x.NumberOfStatusBarOpenPullRequestList);
-            this.showCurrentPullRequestCommand = new UsageTrackingCommand(showCurrentPullRequestCommand,
-                usageTracker, x => x.NumberOfShowCurrentPullRequest);
+            this.openPullRequestsCommand = new UsageTrackingCommand(usageTracker,
+                x => x.NumberOfStatusBarOpenPullRequestList, openPullRequestsCommand);
+            this.showCurrentPullRequestCommand = new UsageTrackingCommand(usageTracker,
+                x => x.NumberOfShowCurrentPullRequest, showCurrentPullRequestCommand);
 
             this.pullRequestSessionManager = pullRequestSessionManager;
             this.teamExplorerContext = teamExplorerContext;
@@ -142,38 +140,6 @@ namespace GitHub.InlineReviews.Services
         {
             var contentControl = mainWindow?.Template?.FindName(StatusBarPartName, mainWindow) as ContentControl;
             return contentControl?.Content as StatusBar;
-        }
-
-        class UsageTrackingCommand : ICommand
-        {
-            readonly ICommand command;
-            readonly Lazy<IUsageTracker> usageTracker;
-            readonly Expression<Func<UsageModel.MeasuresModel, int>> counter;
-
-            internal UsageTrackingCommand(ICommand command, Lazy<IUsageTracker> usageTracker,
-                Expression<Func<UsageModel.MeasuresModel, int>> counter)
-            {
-                this.command = command;
-                this.usageTracker = usageTracker;
-                this.counter = counter;
-            }
-
-            public event EventHandler CanExecuteChanged
-            {
-                add { command.CanExecuteChanged += value; }
-                remove { command.CanExecuteChanged -= value; }
-            }
-
-            public bool CanExecute(object parameter)
-            {
-                return command.CanExecute(parameter);
-            }
-
-            public void Execute(object parameter)
-            {
-                command.Execute(parameter);
-                usageTracker.Value.IncrementCounter(counter).Forget();
-            }
         }
     }
 }
